Public/cm-crawl-master/requirements.txt
imap-tools==0.41.0
schedule==1.1.0
requests==2.25.1
python-dotenv==0.17.1
